% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_extract.R
\name{any_extract}
\alias{any_extract}
\title{any_extract}
\usage{
any_extract(extract_from, extract_at, id_column = NULL)
}
\arguments{
\item{id_column}{Variable in \verb{extract at} that contains an identifier. Default is \code{NULL}. If \code{NULL}, an identifier column will be guessed by searching for column names that contain "ID" or "id" and have as many unique values as the number of rows in \verb{extract at}}
}
\value{

}
\description{
any_extract
}
