% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_import_list.R
\name{read_import_list}
\alias{read_import_list}
\title{read_import_list}
\usage{
read_import_list(filepath)
}
\arguments{
\item{filepath}{filepath to the csv containing the dataset informations}
}
\value{

}
\description{
read_import_list
}
